@dimitri 2024-05-13
Fixes #754